Remove NSApplication dependency for app type
