+++ BrowserServicesKit/Sources/Common/AppVersion.swift
@@ -61,4 +61,46 @@ public struct AppVersion {
         return "\(os.majorVersion).\(os.minorVersion).\(os.patchVersion)"
     }
 
+    public enum AppRunType: String {
+        case normal
+        case unitTests
+        case integrationTests
+        case uiTests
+        case uiTestsOnboarding
+        case xcPreviews
+
+        /// Defines if app run type requires loading full environment, i.e. databases, saved state, keychain etc.
+        public var requiresEnvironment: Bool {
+            switch self {
+            case .normal, .integrationTests, .uiTests, .uiTestsOnboarding:
+                return true
+            case .unitTests, .xcPreviews:
+                return false
+            }
+        }
+    }
+
+    public static let runType: AppRunType = {
+        let isCI = ProcessInfo.processInfo.environment["CI"] != nil
+
+        if let testBundlePath = ProcessInfo().environment["XCTestBundlePath"] {
+            if testBundlePath.contains("Unit") {
+                return .unitTests
+            } else if testBundlePath.contains("Integration") || testBundlePath.contains("DBPE2ETests") {
+                return .integrationTests
+            } else {
+                return .uiTests
+            }
+        } else if ProcessInfo().environment["XCODE_RUNNING_FOR_PREVIEWS"] == "1" {
+            return .xcPreviews
+        } else if ProcessInfo.processInfo.environment["UITEST_MODE_ONBOARDING"] == "1"{
+            return .uiTestsOnboarding
+        } else if ProcessInfo.processInfo.environment["UITEST_MODE"] == "1" || isCI {
+            return .uiTests
+        } else {
+            return .normal
+        }
+    }()
+
+    public var runType: AppRunType { Self.runType }
 }
