+++ macOS/LocalPackages/AppKitExtensions/Sources/AppKitExtensions/NSApplicationExtension.swift
@@ -25,48 +25,6 @@ public extension NSApplication {
         ProcessInfo.processInfo.environment["APP_SANDBOX_CONTAINER_ID"] != nil
     }
 
-    enum RunType: String {
-        case normal
-        case unitTests
-        case integrationTests
-        case uiTests
-        case uiTestsOnboarding
-        case xcPreviews
-
-        /// Defines if app run type requires loading full environment, i.e. databases, saved state, keychain etc.
-        public var requiresEnvironment: Bool {
-            switch self {
-            case .normal, .integrationTests, .uiTests, .uiTestsOnboarding:
-                return true
-            case .unitTests, .xcPreviews:
-                return false
-            }
-        }
-    }
-
-    static let runType: RunType = {
-        let isCI = ProcessInfo.processInfo.environment["CI"] != nil
-
-        if let testBundlePath = ProcessInfo().environment["XCTestBundlePath"] {
-            if testBundlePath.contains("Unit") {
-                return .unitTests
-            } else if testBundlePath.contains("Integration") || testBundlePath.contains("DBPE2ETests") {
-                return .integrationTests
-            } else {
-                return .uiTests
-            }
-        } else if ProcessInfo().environment["XCODE_RUNNING_FOR_PREVIEWS"] == "1" {
-            return .xcPreviews
-        } else if ProcessInfo.processInfo.environment["UITEST_MODE_ONBOARDING"] == "1"{
-            return .uiTestsOnboarding
-        } else if ProcessInfo.processInfo.environment["UITEST_MODE"] == "1" || isCI {
-            return .uiTests
-        } else {
-            return .normal
-        }
-    }()
-    var runType: RunType { Self.runType }
-
     var isCommandPressed: Bool {
         currentEvent?.modifierFlags.contains(.command) ?? false
     }
